requirements.txt
PyAudio
vosk
Requests
playsound
python-dotenv
NetHyTech-STT
groq
edge-tts
googlesearch-python
opencv-python
pvporcupine==1.9.5
pygame
webscout
langdetect
pydub
ffmpeg-python
faster-whisper
googletrans==4.0.0-rc1